refactor(trusty-common): extract slugify_string as single source of truth (closes #1348)
PR #1347 left two byte-for-byte copies of the project-slug rule:
`trusty_memory::messaging::slugify_string` and a private `slugify` in
trusty-controller's `tctl ensure` identity module. Because the controller's
palace name must match the slug trusty-memory's `validate_palace_name`
re-derives, the duplication was a silent-divergence hazard — a tweak to one
copy would let palace creation 400 without any compile-time signal.

This hoists the rule into `trusty-common::slug::slugify_string` (re-exported as
`trusty_common::slugify_string`) as the one source of truth:

- trusty-common: new `slug` module + crate-root re-export; canonical behaviour
  pinned in `slug::tests::slug_derivation_cases`.
- trusty-memory: `messaging::slugify_string` is now a `pub use` shim of the
  trusty-common fn (zero call-site churn); the parity test stays green.
- trusty-controller: deletes the duplicate `slugify` fn and its stale
  "byte-for-byte" comment; `palace_slug` now calls the shared fn directly.

trusty-common 0.15.3 -> 0.16.0 (minor, additive public API). The single pin
point in the root [workspace.dependencies] is updated; the per-crate path-dep
pins in trusty-memory / trusty-search / trusty-gworkspace are advanced to match
so workspace resolution and future publish stay consistent.
