+++ crates/trusty-common/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "trusty-common"
-version = "0.15.3"
+version = "0.16.0"
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
